refactor(api): 重构用户管理API和前端日期显示组件
- 将用户详情API从混合方法文件分离为独立的GET请求处理器
- 更新README文档中的API路由文件命名约定
- 重构ScheduleList组件中的日期格式化和显示逻辑
- 优化2FA验证中的限流机制，区分用户和IP维度的失败计数
